prov/tcp2: Move protocol definitions to own header file
Isolate protocol from node local structures.
